+++ libraries/geomath/test/unit_test_001.cpp
@@ -40,9 +40,9 @@ unittest_teardown()
 
 unittest(test_constants)
 {
-  assertEqualFloat(GRAVITY_STANDARD, 9.80655, 0.001);
-  assertEqualFloat(GRAVITY_EQUATOR,  9.78033, 0.001);
-  assertEqualFloat(GRAVITY_POLES,    9.8337,  0.001);
+  assertEqualFloat(GRAVITY_STANDARD, 9.80655,  0.00001);
+  assertEqualFloat(GRAVITY_EQUATOR,  9.780327, 0.00001);
+  assertEqualFloat(GRAVITY_POLES,    9.832185, 0.00001);
 }